AI/inference/requirements.txt
fastapi
uvicorn[standard]
tensorflow==2.14.0
pillow
numpy
python-multipart
